Web_Research/requirements.txt
streamlit==1.25.0
langchain==0.0.244
chromadb==0.4.3
openai==0.27.8
html2text==2020.1.16
tiktoken==0.4.0
faiss-cpu==1.7.4